Add attributes filter for node endpoints
By default we are not returning any more attributes and extras
in nodes list and in node details.

If attributes is passed in the query string and is set to True,
then all attributes are listed.
The attributes can also be a comma-separated list of attribute
keys, and in this case only the specified attributes keys
will be returned in the response.
For nodes that do not have the specified attributes keys, they
will list the attribute keys and their value will be null.

The nodes/<id> endpoint follow the same behavior.