Cleaned tree endpoint, modified attributes and extras filters
The /nodes/<id>/links/tree endpoint returns a nodes
list, containing a dict with the information relative
to the main node.
Among the keys of this dict, incoming refers to a
list of incoming nodes (one dict per node) and outgoing
to a list of outgoing nodes.
The fields in the dicts of incoming and outgoing lists
are the same returned by the /nodes/<id>/links/incoming
and /nodes/<id>/links/outgoing endpoints.

Regarding the attributes and extras filters for the
/nodes endpoint, specifying in the query string
attributes=true one gets all attributes, while to get
specific attributes one has to set in the query string
attributes=true and attributes_filter = list of
comma-separated values of attributes on which to project.
The extras and extras_filter filters follow the same
behavior.