Cleaned /nodes/<id>/download endpoint
Moved the logic to get data to download to the
get_downloadable_data method in the DataTranslator
class.
The user can specify the download format using the
download_format parameter in the query string.
The download parameter is used to specify if the
data has to be returned in the json response
(by setting download=False) or as download file.

The method is not implemented for Process Nodes.